backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
psycopg2-binary==2.9.9
drf-yasg==1.21.7
setuptools>=65.0.0
gunicorn==21.2.0
whitenoise==6.6.0

